+++ client/src/components/layout/Sidebar.vue
@@ -46,7 +46,7 @@
     </nav>
 
     <div class="p-4 border-t border-gray-800 text-sm text-gray-400">
-      <div>v0.30.2-alpha</div>
+      <div>v0.30.3-alpha</div>
     </div>
   </aside>
 </template>